Move ListFieldInput into Opal
A raw input with no app dependencies — it composes InputTypeIn and Tag,
both already Opal. Moves to opal/components/inputs/list-field-input with
the house structure; its two consumers import from the barrel.

Kept distinct from InputTags deliberately: this renders its chips below
the field rather than inline within it.
